Gathering View
Added popup and associated logic for the view buttons for the gatherings. I also fixed some stuff for the update popup that would prevent it from working correctly. none of this has been tested yet, so I am not sure if there are any major issues with the code. Will test as soon as possible.